% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocompute-package.R
\docType{package}
\name{biocompute-package}
\alias{biocompute}
\alias{biocompute-package}
\title{biocompute: Create and Manipulate BioCompute Objects}
\description{
Tools to create, validate, and export BioCompute Objects
    described in King et al. (2019) <doi:10.17605/osf.io/h59uh>.
    Users can encode information in data frames, and compose
    BioCompute Objects from the domains defined by the standard.
    A checksum validator and a JSON schema validator are provided.
    This package also supports exporting BioCompute Objects as JSON,
    PDF, HTML, or 'Word' documents, and exporting to cloud-based platforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sbg.github.io/biocompute/}
  \item \url{https://github.com/sbg/biocompute}
  \item Report bugs at \url{https://github.com/sbg/biocompute/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao@sevenbridges.com} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Soner Koc \email{soner.koc@sevenbridges.com}
  \item Dennis Dean \email{dennis.dean@sevenbridges.com} (\href{https://orcid.org/0000-0002-7621-9717}{ORCID})
}

Other contributors:
\itemize{
  \item Seven Bridges Genomics [copyright holder, funder]
}

}
\keyword{internal}
