% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_v1.4.2}
\alias{compose_v1.4.2}
\alias{compose}
\title{Compose BioCompute Object (v1.4.2)}
\usage{
compose_v1.4.2(
  tlf,
  provenance,
  usability,
  extension,
  description,
  execution,
  parametric,
  io,
  error
)

compose(
  tlf,
  provenance,
  usability,
  extension,
  description,
  execution,
  parametric,
  io,
  error
)
}
\arguments{
\item{tlf}{Top level fields}

\item{provenance}{Provenance domain}

\item{usability}{Usability domain}

\item{extension}{Extension domain}

\item{description}{Description domain}

\item{execution}{Execution domain}

\item{parametric}{Parametric domain}

\item{io}{I/O domain}

\item{error}{Error domain}
}
\value{
A list of class \code{bco}
}
\description{
Compose BioCompute Object (v1.4.2)
}
\examples{
tlf <- compose_tlf(
  compose_provenance(), compose_usability(), compose_extension(),
  compose_description(), compose_execution(), compose_parametric(),
  compose_io(), compose_error()
)
biocompute::compose(
  tlf,
  compose_provenance(), compose_usability(), compose_extension(),
  compose_description(), compose_execution(), compose_parametric(),
  compose_io(), compose_error()
) \%>\% convert_json()
}
