% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_yaml}
\alias{convert_yaml}
\title{Convert BioCompute Object or domain to YAML string}
\usage{
convert_yaml(x, ...)
}
\arguments{
\item{x}{BioCompute Object or domain}

\item{...}{Additional parameters for \code{\link[yaml]{as.yaml}}.}
}
\value{
YAML string of the BioCompute Object
}
\description{
Convert BioCompute Object or domain to YAML string
}
\examples{
compose_description() \%>\%
  convert_yaml() \%>\%
  cat()
generate_example("minimal") \%>\%
  convert_yaml() \%>\%
  cat()
}
