% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-json.R
\name{export_json}
\alias{export_json}
\title{Export BioCompute Object as JSON}
\usage{
export_json(x, file)
}
\arguments{
\item{x}{BioCompute Object JSON string from \code{\link{convert_json}}}

\item{file}{JSON file path}
}
\description{
Export BioCompute Object as JSON
}
\examples{
bco <- tempfile(fileext = ".json")
generate_example("HCV1a") \%>\%
  convert_json() \%>\%
  export_json(bco)
cat(paste(readLines(bco), collapse = "\\n"))
}
