% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compose.R
\name{generate_example}
\alias{generate_example}
\title{Generate example BioCompute Objects}
\usage{
generate_example(type = c("minimal", "HCV1a"))
}
\arguments{
\item{type}{Example type. Default is \code{"minimal"}.}
}
\value{
Example BioCompute Object
}
\description{
Generate example BioCompute Objects
}
\examples{
generate_example("minimal") \%>\% convert_json()
}
