% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compose.R
\name{generate_id}
\alias{generate_id}
\title{Generate ID for the BioCompute Object}
\usage{
generate_id(platform = c("sevenbridges"))
}
\arguments{
\item{platform}{Platform. Default is \code{"sevenbridges"}.}
}
\value{
BioCompute Object ID
}
\description{
Generate ID for the BioCompute Object
}
\examples{
generate_id()
}
