% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compose.R
\name{is_bco}
\alias{is_bco}
\title{Is this a BCO object?}
\usage{
is_bco(x)
}
\arguments{
\item{x}{any object}
}
\value{
Logical. \code{TRUE} if it is a BCO object, \code{FALSE} if not.
}
\description{
Is this a BCO object?
}
\examples{
generate_example("minimal") \%>\% is_bco()
}
