% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compose.R
\name{is_domain}
\alias{is_domain}
\title{Is this a domain object?}
\usage{
is_domain(x)
}
\arguments{
\item{x}{any object}
}
\value{
Logical. \code{TRUE} if it is a domain object, \code{FALSE} if not.
}
\description{
Is this a domain object?
}
\examples{
is_domain(compose_description())
}
