% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bco.R
\name{read_bco}
\alias{read_bco}
\title{Parse Biocompute Object From JSON File to R Object}
\usage{
read_bco(x, ...)
}
\arguments{
\item{x}{BioCompute Object .json file}

\item{...}{Additional parameters for \code{\link[jsonlite]{fromJSON}}.}
}
\value{
A list of class \code{bco}
}
\description{
Parse Biocompute Object From JSON File to R Object
}
\examples{
bco <- tempfile(fileext = ".json")
bco <- generate_example("HCV1a") \%>\%
  convert_json() \%>\%
  export_json(bco)
bco \%>\% read_bco()
}
