% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{versions}
\alias{versions}
\title{BioCompute Object specification versions}
\usage{
versions()
}
\value{
List of current and all available BioCompute Object
specification versions supported by the package.
}
\description{
BioCompute Object specification versions
}
\examples{
versions()
}
